loss fix

+++ luxonis_train/nodes/backbones/recsubnet/recsubnet.py
@@ -38,7 +38,23 @@ class RecSubNet(BaseNode[Tensor, Tuple[Tensor, Tensor, Tensor]]):
             else:
                 h, w = x.shape[-2:]
                 an_mask = torch.zeros((x.shape[0], h, w), device=x.device)
-            an_mask = an_mask.unsqueeze(1)
+            an_mask = an_mask.unsqueeze(1).long()
+
+        # if self.training:
+        #     import matplotlib.pyplot as plt
+
+        #     # mask is (batch, h, w)
+        #     mask = an_mask[0][0].cpu().numpy()
+        #     plt.imshow(mask)
+        #     plt.show(block=True)
+
+        #     # imagenet denorm on x[0]
+        #     x_plt = x[0].cpu().numpy()
+        #     x_plt = x_plt.transpose(1, 2, 0)
+        #     x_plt = x_plt * 0.229 + 0.485
+        #     plt.imshow(x_plt)
+        #     plt.show(block=True)
+
         b5 = self.encoder(x)
 
         output = self.decoder(b5)
